Use Mesos network bandwidth for Spark executors
At criteo, containers are network isolated and if no network bandwidth
resource is allocated, Mesos will reserve a default amount. This
patch provide a way for user to specify the amount of network bandwidth
required by the executors and therefore used by the tasks.

This fix is Criteo specific as network_bandwidth is a custom resource
we introduced in our clusters and the code of the isolator is not
public (yet).