[CRITEO] Fix SHS performance (#133)
SPARK-38896 introduce a huge performance regression. At many places, instead of manipulating an iterator on the database data and stopping eagerly, SHS now constructs an intermediate list of the entire SHS database data and then returns an iterator on it, which causes sever performance degradation.

We introduce a new method to transform directly the iterator within the boundary of the try-with-resources introduced with SPARK-38896. This allows to keep the expected KVStoreIterator recycling, while allowing to compute the minimal information required.

The perfomance boost goes from 30s+ to load the main SHS page to less than 2s.

Co-authored-by: w.montaz <w.montaz@criteo.com>