[CRITEO] fix key issue while counting rows (#164)
When hiveSerde is used (spark.sql.hive.convertInsertingPartitionedTable=false),
Say that dataset has hour as partition , hour is int and should be formatted in two digit format (hour=06) stored as int with padding inside metastore but rows inside the tables have it as int without the padding.
when Analyze Table .. partition(hour='06') compute statistics spark use the info inside the rows to generate row count of the given partition since hour is store without padding inside the table , so as spark infer rows to determine the partition key and it rows count it creates partition key that it is different then the one given in the input (that reflect what we have in metastore). Thus the created stats are never saved in metastore due to this format mismatch.

When hiveSerde is not used we would like to keep the logic as it is since this use case will not happen